requirements.txt
azure-identity==1.15.0
msal==1.24.1
msal-extensions==1.0.0
openai==1.1.1 
python-dotenv==1.0.0